fix: fixed cant add the firstApplication Key
In the OctoPrint Settings page, on the Application Keys feature,
you couldn't add any new key when there are no keys present.

After this change, it will show `Manually generate an application key`
text and it will allow you to create a new application key.

The way to replicate is, on a local OctoPrint

```bash
rm -rf ~/.octoprint
octoprint serve
```

Go in a browser to http://127.0.0.1:5000 and after configuration,
navigate to the `Application Keys` page.

Not sure why this one is not allowing, while the `template_usersettings`
has the fix.